+++ BBGE/ActionSet.h
@@ -60,7 +60,7 @@ public:
 	ActionInput *getActionInputByName(const std::string &name);
 
 	InputDevice inputMode;
-	size_t joystickID; // >= 0: use that, -1 = no joystick, or ACTIONSET_REASSIGN_JOYSTICK
+	int joystickID; // >= 0: use that, -1 = no joystick, or ACTIONSET_REASSIGN_JOYSTICK
 
 	// --- Saved in config ---
 	ActionInputSet inputSet;
